Consolidate all test files under src/__tests__
Move player.test.tsx, use-simple-jack.test.tsx, get-hand-score.test.ts, and mock-deck-generator.test.ts from their co-located positions into src/__tests__. Update relative imports accordingly.
